somehow replaced old spi component
