+++ components/route/include/route.h
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "spi_tx_netif.h"
+#include "spi_netif.h"
 
 #define MAX_NUMBER_ROUTES 10
