+++ main/test_board.c
@@ -4,14 +4,14 @@ const char *BOARD_TEST_TAG = "==> BOARD TEST";
 
 void send_test_payload(){
     ESP_LOGI(BOARD_TEST_TAG, "sending %s", SPI_TEST_PAYLOAD);
-    spi_transmit(SPI_TEST_PAYLOAD, sizeof(SPI_TEST_PAYLOAD));
+    spi_lowlevel_transmit(SPI_TEST_PAYLOAD, sizeof(SPI_TEST_PAYLOAD));
 }
 
 void receive_test_payload(){
-    char test_buffer[SPI_PAYLOAD_BUFFER_SIZE];
+    char test_buffer[sizeof(SPI_TEST_PAYLOAD)];
     size_t received;
 
-    received = spi_receive(test_buffer);
+    received = spi_lowlevel_receive(test_buffer, sizeof(SPI_TEST_PAYLOAD));
     printf("%s, %d\n", test_buffer, received);
     assert(strcmp(test_buffer, SPI_TEST_PAYLOAD) == 0);
     ESP_LOGI(BOARD_TEST_TAG, "received %s", test_buffer);
